+++ NEWS.md
@@ -3,7 +3,12 @@
 - `score_forecasts()` is now implemented.
 
   Minimally one should have the `forecasts` dataframe produced by `load_forecasts()` and the truth dataframe produced by `load_truth()` to calculate scores. If one desires to specify a subset of all available scores, one should consult [this reference](https://epiforecasts.io/scoringutils/reference/eval_forecasts.html#details) for valid scores in the `desired_score_types` vector.
+
+- Update `plot_forecast()`
   
+  Set `truth_source` to be optional when the user provides `truth_data`. However, it is still needed when `show_caption = TRUE`.
+  
+  Remove format validation for `model` column in user-provided `truth_data`.
 
 ## covidHubUtils 0.1.1
 
@@ -35,8 +40,9 @@ This is the first version of the package with a 0.x release.
 
 ### v 0.1
  - initial release: big plotting changes
- - v0.1.1 :
+ - v0.1.1:
     - in load_latest_forecasts, use provided source for calls to lower level functions
  - v0.1.2:
     - support daily hospitalization plot in plot_forecast
     - add `facet_nrow`, `facet_ncol`, and `fill_transparency` to `plot_forecast` function
+